write to file only if build_dir-config present

+++ potypo/__main__.py
@@ -44,19 +44,26 @@ def main():
         filter_list.append(filters.make_EdgecaseFilter(words))
 
 
-    def errmsg(outputfile, path, linenum, word):
+    def errmsg(outputfile, path, linenum, word, write_file):
         print("ERROR: {}:{}: {}".format(path, linenum, word))
-        outputfile.write("ERROR: {}:{}: {}\n".format(path, linenum, word))
+        if write_file:
+            outputfile.write("ERROR: {}:{}: {}\n".format(path, linenum, word))
 
     try:
+        write_file = conf['build_dir']
+    except KeyError:
+        write_file = None
+
+    if write_file:
         print('Creating build directory at', conf['build_dir'])
-        os.mkdir(conf['build_dir'])
-    except FileExistsError:
-        print("File or directory", conf['build_dir'], "already exists, deleting")
-        rmtree(conf['build_dir'])
-        print('Recreating build directory')
-        os.mkdir(conf['build_dir'])
-        print('Build directory created')
+        try:
+            os.mkdir(conf['build_dir'])
+        except FileExistsError:
+            print("File or directory", conf['build_dir'], "already exists, deleting")
+            rmtree(conf['build_dir'])
+            print('Recreating build directory')
+            os.mkdir(conf['build_dir'])
+            print('Build directory created')
 
     # checks contains one Check-Object for every po-file
     checks = []
@@ -64,12 +71,13 @@ def main():
     for root, dirs, files in os.walk(conf['locales_dir']):
         for f in files:
             if f.endswith(".po"):
-                checks.append(Check(os.path.join(root, f), conf['build_dir'], conf['ignores_dir'], chunker_list, filter_list))
+                checks.append(Check(os.path.join(root, f), write_file, conf['ignores_dir'], chunker_list, filter_list))
 
     en_ignorefile = Check.get_ignorefile(conf['default_language'], conf['ignores_dir'])
     en_dict = DictWithPWL(conf['default_language'], pwl=en_ignorefile)
     en_ckr = SpellChecker(en_dict, chunkers=chunker_list, filters=filter_list)
-    output_file = open(os.path.join(conf['build_dir'], 'en_output.txt'), 'w')
+    if write_file:
+        output_file = open(os.path.join(write_file, 'en_output.txt'), 'w')
 
     for c in checks:
         for entry in c.po:
@@ -79,14 +87,17 @@ def main():
             en_ckr.set_text(entry.msgid)
             for err in en_ckr:
                 path = os.path.relpath(c.popath, start=config['potypo']['locales_dir'])
-                errmsg(output_file, path, entry.linenum, err.word)
+                errmsg(output_file, path, entry.linenum, err.word, write_file)
 
             c.checker.set_text(entry.msgstr)
             for err in c.checker:
                 path = os.path.relpath(c.popath, start=config['potypo']['locales_dir'])
-                errmsg(c.output_file, path, entry.linenum, err.word)
+                errmsg(c.output_file, path, entry.linenum, err.word, write_file)
 
-    print("Spell-checking done. You can find the outputs in", config['potypo']['build_dir'] + "/<lang>/{js_}output")
+    if write_file:
+        print("Spell-checking done. You can find the outputs in", write_file + "/<lang>/{js_}output")
+    else:
+        print("Spell-checking done.")
 
 if __name__ == "__main__":
     main()
